requirements.txt
streamlit
tensorflow-cpu==2.15.0
opencv-python
mediapipe==0.10.9
scikit-learn
numpy
pandas
scipy
scikit-image
matplotlib
openpyxl
opendatasets
